fix(version): remove whitespace at end of APP_VERSION constant
The Dockerfile contains `echo $VERSION > /version.txt`, where `echo`
implicitly causes a newline at the end of the file. To avoid that,
one could use `echo -n`, but then the file isn't nicely readable when
doing a `cat` or something simular.

This fix `rtrims` (i.e. "Strip whitespace (or other
characters) from the end of a string") the contents of the version file.

Fixes kanboard/kanboard#4708