config.php.default: clarify meaning of LDAP_USER_CREATION
While the documentation explains the meaning of LDAP_USER_CREATION very well, the example config file just contains the explanation "Allow automatic LDAP user creation". This might confuse some admins (such as me) as it can be interpreted as "create an LDAP user for new local users". This commit expands the explanation to clarify the meaning of LDAP_USER_CREATION.